make the public api use only arithmetic expression
Now, qeval never returns a dict but always an arithmetic expression for
quantity and quantity points.
Quantities are of the form Value*Quantity[Unit] and quantity points of
the form Origin + Value*Quantity[Unit].
If Origin is a variable, it needs to be prefixed by o:Origin to be
correctly parsed.